Create The Association Between Heart Disease and Heavy Alcohol Drinking Among Adults in the United States
Heart disease (also called cardiovascular disease) refers to the various types of problems of the heart - the disease of heart and blood vessels. It is the leading cause of death in the United States, affecting most racial groups and genders (CDC, 2023). The objective of this study is to explore the association between heart disease and heavy alcohol drinking, while controlling for one’s age, sex, exercise, race, smoking and general health.